# perf(core)!: optimise interop with crypto wasm