--- a/packages/core/crypto/util.ts
+++ b/packages/core/crypto/util.ts
@@ -1,35 +1,36 @@
 import { decodeHex, encodeHex } from '@std/encoding/hex'
-import type { Bytes as WasmBytes } from './wasm/iroha_crypto_wasm.js'
+
+export type BytesRepr = { t: 'array'; array: Uint8Array } | { t: 'hex'; hex: string }
 
 /**
  * Helper to work with binary data passed into the WASM
  */
 export class Bytes {
-  public static array(data: Uint8Array): Bytes {
-    return new Bytes({ t: 'array', c: data })
+  public static array(array: Uint8Array): Bytes {
+    return new Bytes({ t: 'array', array })
   }
 
   public static hex(hex: string): Bytes {
-    return new Bytes({ t: 'hex', c: hex })
+    return new Bytes({ t: 'hex', hex })
   }
 
-  #data: WasmBytes
+  #repr: BytesRepr
 
-  private constructor(data: WasmBytes) {
-    this.#data = data
+  private constructor(data: BytesRepr) {
+    this.#repr = data
   }
 
-  public get asWasmFormat(): WasmBytes {
-    return this.#data
+  public get repr(): BytesRepr {
+    return this.#repr
   }
 
   public hex(): string {
-    if (this.#data.t == 'hex') return this.#data.c
-    return encodeHex(this.#data.c)
+    if (this.#repr.t == 'hex') return this.#repr.hex
+    return encodeHex(this.#repr.array)
   }
 
   public array(): Uint8Array {
-    if (this.#data.t === 'array') return this.#data.c
-    return decodeHex(this.#data.c)
+    if (this.#repr.t === 'array') return this.#repr.array
+    return decodeHex(this.#repr.hex)
   }
 }